GlusterFS: Set permissions on qcow2 snapshot files
The GlusterFS driver sets specific permissions on the
volume file when volumes are created.  This is required
for volume attaches to work when using QEMU's libgfapi
until a libvirt change is landed to enable libvirt to set
the permissions itself. [1]

This does not weaken the current security model in any
significant way, as it only lines up the permissions used
for snapshot files with how volume files are already created.

This enables libgfapi volume attaches to work correctly when
snapshots exist.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1052014

Change-Id: I93944724c557c907eb3821a061e7434f09ff20bc