RBD: Flattening of child volumes during deletion
If a volume has clone volumes (also known as child images), Cinder
should flatten the child volumes when deleting an RBD parent volume to
prevent the delete from failing due to ImageBusy.

The intent is to only check for child images when we cannot
successfully perform the original remove(), as this is a quick way to do
it.

A new configuration option, rbd_concurrent_flatten_operations, was
introduced to limit how many flatten calls can be made at the same time.
This is to prevent overloading the backend. The default value is 3.

Co-Author-By: Eric Harney <eharney@redhat.com>
Co-Author-By: Sofia Enriquez <lsofia.enriquez@gmail.com>

Closes-Bug: #1969643
Change-Id: I009d0748fdc829ca0b4f99bc9b70dadd19717d04