NFS: Fix cloning encrypted volumes
This patch adds code to the _copy_volume_from_snapshot method to enable
cloning of encrypted NFS volumes. Currently, when attempting to clone a
nfs volume, a snapshot of the volume is created and then converted into
a volume.

In particular, when using encryption, qemu requires the backingfile with
the encryption key in order to convert the snapshot into an independent
volume.

Closes-Bug: #2007139
Change-Id: I8070aa4829ae55efd648325558be50a0b9a10f8c